+++ jme3-plugins/src/fbx/java/com/jme3/scene/plugins/fbx/file/FbxDump.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009-2015 jMonkeyEngine
+ * Copyright (c) 2009-2020 jMonkeyEngine
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -148,9 +148,9 @@ public final class FbxDump {
             case 'F':
                 // Double, Float.
                 if (property instanceof Double) {
-                    ps.print(DECIMAL_FORMAT.format((Double)property));
+                    ps.print(DECIMAL_FORMAT.format(property));
                 } else if (property instanceof Float) {
-                    ps.print(DECIMAL_FORMAT.format((Float)property));
+                    ps.print(DECIMAL_FORMAT.format(property));
                 } else {
                     ps.print(property);
                 }
